src/web/requirements.txt
python-dotenv
requests
flask
